Front page construction and DB
Beginning the construction of the front page, finally got the header to
work, only one aspect left to fix which is to align account, basket and
login button to the right, currently left. Also in the process of
sorting out the domain too. Create and sourced some new images and
implementing retrieval from DB