Comments for the previous commit
In terms of the producer page, I was trying to intergrate Google Maps
api to hardcode my own map, I was having a little trouble with it, again
this is something I want to come back to and is currently commented out.

Lastly the test files are a test scenario showing the use of AJAX in
conjunction with PHP and mySQL to retrieve data from the database
without the to reload the page. This example works showing results, I am
now moving onto another test scenario to have the products already in
place and filter them....we'll see how it goes